Replaced size_t by unsigned int

+++ src/external/tinyobj_loader_c.h
@@ -24,8 +24,8 @@
 #ifndef TINOBJ_LOADER_C_H_
 #define TINOBJ_LOADER_C_H_
 
-/* @todo { Remove stddef dependency. size_t? } */
-#include <stddef.h>
+/* @todo { Remove stddef dependency. unsigned int? } ---> RAY: DONE. */
+//#include <stddef.h>
 
 typedef struct {
   char *name;
@@ -93,18 +93,18 @@ typedef struct {
  * Returns TINYOBJ_ERR_*** when there is an error.
  */
 extern int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
-                             size_t *num_shapes, tinyobj_material_t **materials,
-                             size_t *num_materials, const char *buf, size_t len,
+                             unsigned int *num_shapes, tinyobj_material_t **materials,
+                             unsigned int *num_materials, const char *buf, unsigned int len,
                              unsigned int flags);
 extern int tinyobj_parse_mtl_file(tinyobj_material_t **materials_out,
-                                  size_t *num_materials_out,
+                                  unsigned int *num_materials_out,
                                   const char *filename);
 
 extern void tinyobj_attrib_init(tinyobj_attrib_t *attrib);
 extern void tinyobj_attrib_free(tinyobj_attrib_t *attrib);
-extern void tinyobj_shapes_free(tinyobj_shape_t *shapes, size_t num_shapes);
+extern void tinyobj_shapes_free(tinyobj_shape_t *shapes, unsigned int num_shapes);
 extern void tinyobj_materials_free(tinyobj_material_t *materials,
-                                   size_t num_materials);
+                                   unsigned int num_materials);
 
 #ifdef TINYOBJ_LOADER_C_IMPLEMENTATION
 #include <stdio.h>
@@ -155,8 +155,8 @@ static int until_space(const char *token) {
   return (int)(p - token);
 }
 
-static size_t length_until_newline(const char *token, size_t n) {
-  size_t len = 0;
+static unsigned int length_until_newline(const char *token, unsigned int n) {
+  unsigned int len = 0;
 
   /* Assume token[n-1] = '\0' */
   for (len = 0; len < n - 1; len++) {
@@ -171,8 +171,8 @@ static size_t length_until_newline(const char *token, size_t n) {
   return len;
 }
 
-static size_t length_until_line_feed(const char *token, size_t n) {
-  size_t len = 0;
+static unsigned int length_until_line_feed(const char *token, unsigned int n) {
+  unsigned int len = 0;
 
   /* Assume token[n-1] = '\0' */
   for (len = 0; len < n; len++) {
@@ -202,7 +202,7 @@ static int my_atoi(const char *c) {
 }
 
 /* Make index zero-base, and also support relative index. */
-static int fixIndex(int idx, size_t n) {
+static int fixIndex(int idx, unsigned int n) {
   if (idx > 0) return idx - 1;
   if (idx == 0) return 0;
   return (int)n + idx; /* negative value = relative */
@@ -453,9 +453,9 @@ static void parseFloat3(float *x, float *y, float *z, const char **token) {
   (*z) = parseFloat(token);
 }
 
-static char *my_strdup(const char *s, size_t max_length) {
+static char *my_strdup(const char *s, unsigned int max_length) {
   char *d;
-  size_t len;
+  unsigned int len;
 
   if (s == NULL) return NULL;
 
@@ -465,15 +465,15 @@ static char *my_strdup(const char *s, size_t max_length) {
 
   /* trim line ending and append '\0' */
   d = (char *)TINYOBJ_MALLOC(len + 1); /* + '\0' */
-  memcpy(d, s, (size_t)(len));
+  memcpy(d, s, (unsigned int)(len));
   d[len] = '\0';
 
   return d;
 }
 
-static char *my_strndup(const char *s, size_t len) {
+static char *my_strndup(const char *s, unsigned int len) {
   char *d;
-  size_t slen;
+  unsigned int slen;
 
   if (s == NULL) return NULL;
   if (len == 0) return NULL;
@@ -491,10 +491,10 @@ static char *my_strndup(const char *s, size_t len) {
   return d;
 }
 
-char *dynamic_fgets(char **buf, size_t *size, FILE *file) {
+char *dynamic_fgets(char **buf, unsigned int *size, FILE *file) {
   char *offset;
   char *ret;
-  size_t old_size;
+  unsigned int old_size;
 
   if (!(ret = fgets(*buf, (int)*size, file))) {
     return ret;
@@ -560,8 +560,8 @@ typedef struct
 {
   unsigned long* hashes;
   hash_table_entry_t* entries;
-  size_t capacity;
-  size_t n;
+  unsigned int capacity;
+  unsigned int n;
 } hash_table_t;
 
 static unsigned long hash_djb2(const unsigned char* str)
@@ -576,7 +576,7 @@ static unsigned long hash_djb2(const unsigned char* str)
   return hash;
 }
 
-static void create_hash_table(size_t start_capacity, hash_table_t* hash_table)
+static void create_hash_table(unsigned int start_capacity, hash_table_t* hash_table)
 {
   if (start_capacity < 1)
     start_capacity = HASH_TABLE_DEFAULT_SIZE;
@@ -596,10 +596,10 @@ static void destroy_hash_table(hash_table_t* hash_table)
 static int hash_table_insert_value(unsigned long hash, long value, hash_table_t* hash_table)
 {
   /* Insert value */
-  size_t start_index = hash % hash_table->capacity;
-  size_t index = start_index;
+  unsigned int start_index = hash % hash_table->capacity;
+  unsigned int index = start_index;
   hash_table_entry_t* start_entry = hash_table->entries + start_index;
-  size_t i;
+  unsigned int i;
   hash_table_entry_t* entry;
 
   for (i = 1; hash_table->entries[index].filled; i++)
@@ -648,11 +648,11 @@ static hash_table_entry_t* hash_table_find(unsigned long hash, hash_table_t* has
   return NULL;
 }
 
-static void hash_table_maybe_grow(size_t new_n, hash_table_t* hash_table)
+static void hash_table_maybe_grow(unsigned int new_n, hash_table_t* hash_table)
 {
-  size_t new_capacity;
+  unsigned int new_capacity;
   hash_table_t new_hash_table;
-  size_t i;
+  unsigned int i;
 
   if (new_n <= hash_table->capacity) {
     return;
@@ -680,7 +680,7 @@ static int hash_table_exists(const char* name, hash_table_t* hash_table)
   return hash_table_find(hash_djb2((const unsigned char*)name), hash_table) != NULL;
 }
 
-static void hash_table_set(const char* name, size_t val, hash_table_t* hash_table)
+static void hash_table_set(const char* name, unsigned int val, hash_table_t* hash_table)
 {
   /* Hash name */
   unsigned long hash = hash_djb2((const unsigned char *)name);
@@ -709,7 +709,7 @@ static long hash_table_get(const char* name, hash_table_t* hash_table)
 }
 
 static tinyobj_material_t *tinyobj_material_add(tinyobj_material_t *prev,
-                                                size_t num_materials,
+                                                unsigned int num_materials,
                                                 tinyobj_material_t *new_mat) {
   tinyobj_material_t *dst;
   dst = (tinyobj_material_t *)TINYOBJ_REALLOC(
@@ -720,14 +720,14 @@ static tinyobj_material_t *tinyobj_material_add(tinyobj_material_t *prev,
 }
 
 static int tinyobj_parse_and_index_mtl_file(tinyobj_material_t **materials_out,
-                                            size_t *num_materials_out,
+                                            unsigned int *num_materials_out,
                                             const char *filename,
                                             hash_table_t* material_table) {
   tinyobj_material_t material;
-  size_t buffer_size = 128;
+  unsigned int buffer_size = 128;
   char *linebuf;
   FILE *fp;
-  size_t num_materials = 0;
+  unsigned int num_materials = 0;
   tinyobj_material_t *materials = NULL;
   int has_previous_material = 0;
   const char *line_end = NULL;
@@ -788,7 +788,7 @@ static int tinyobj_parse_and_index_mtl_file(tinyobj_material_t **materials_out,
 #else
       sscanf(token, "%s", namebuf);
 #endif
-      material.name = my_strdup(namebuf, (size_t) (line_end - token));
+      material.name = my_strdup(namebuf, (unsigned int) (line_end - token));
 
       /* Add material to material table */
       if (material_table)
@@ -889,56 +889,56 @@ static int tinyobj_parse_and_index_mtl_file(tinyobj_material_t **materials_out,
     /* ambient texture */
     if ((0 == strncmp(token, "map_Ka", 6)) && IS_SPACE(token[6])) {
       token += 7;
-      material.ambient_texname = my_strdup(token, (size_t) (line_end - token));
+      material.ambient_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* diffuse texture */
     if ((0 == strncmp(token, "map_Kd", 6)) && IS_SPACE(token[6])) {
       token += 7;
-      material.diffuse_texname = my_strdup(token, (size_t) (line_end - token));
+      material.diffuse_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* specular texture */
     if ((0 == strncmp(token, "map_Ks", 6)) && IS_SPACE(token[6])) {
       token += 7;
-      material.specular_texname = my_strdup(token, (size_t) (line_end - token));
+      material.specular_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* specular highlight texture */
     if ((0 == strncmp(token, "map_Ns", 6)) && IS_SPACE(token[6])) {
       token += 7;
-      material.specular_highlight_texname = my_strdup(token, (size_t) (line_end - token));
+      material.specular_highlight_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* bump texture */
     if ((0 == strncmp(token, "map_bump", 8)) && IS_SPACE(token[8])) {
       token += 9;
-      material.bump_texname = my_strdup(token, (size_t) (line_end - token));
+      material.bump_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* alpha texture */
     if ((0 == strncmp(token, "map_d", 5)) && IS_SPACE(token[5])) {
       token += 6;
-      material.alpha_texname = my_strdup(token, (size_t) (line_end - token));
+      material.alpha_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* bump texture */
     if ((0 == strncmp(token, "bump", 4)) && IS_SPACE(token[4])) {
       token += 5;
-      material.bump_texname = my_strdup(token, (size_t) (line_end - token));
+      material.bump_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
     /* displacement texture */
     if ((0 == strncmp(token, "disp", 4)) && IS_SPACE(token[4])) {
       token += 5;
-      material.displacement_texname = my_strdup(token, (size_t) (line_end - token));
+      material.displacement_texname = my_strdup(token, (unsigned int) (line_end - token));
       continue;
     }
 
@@ -962,7 +962,7 @@ static int tinyobj_parse_and_index_mtl_file(tinyobj_material_t **materials_out,
 }
 
 int tinyobj_parse_mtl_file(tinyobj_material_t **materials_out,
-                           size_t *num_materials_out,
+                           unsigned int *num_materials_out,
                            const char *filename) {
   return tinyobj_parse_and_index_mtl_file(materials_out, num_materials_out, filename, NULL);
 } 
@@ -988,10 +988,10 @@ typedef struct {
 
   /* @todo { Use dynamic array } */
   tinyobj_vertex_index_t f[TINYOBJ_MAX_FACES_PER_F_LINE];
-  size_t num_f;
+  unsigned int num_f;
 
   int f_num_verts[TINYOBJ_MAX_FACES_PER_F_LINE];
-  size_t num_f_num_verts;
+  unsigned int num_f_num_verts;
 
   const char *group_name;
   unsigned int group_name_len;
@@ -1011,7 +1011,7 @@ typedef struct {
   CommandType type;
 } Command;
 
-static int parseLine(Command *command, const char *p, size_t p_len,
+static int parseLine(Command *command, const char *p, unsigned int p_len,
                      int triangulate) {
   char linebuf[4096];
   const char *token;
@@ -1073,7 +1073,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
 
   /* face */
   if (token[0] == 'f' && IS_SPACE((token[1]))) {
-    size_t num_f = 0;
+    unsigned int num_f = 0;
 
     tinyobj_vertex_index_t f[TINYOBJ_MAX_FACES_PER_F_LINE];
     token += 2;
@@ -1090,8 +1090,8 @@ static int parseLine(Command *command, const char *p, size_t p_len,
     command->type = COMMAND_F;
 
     if (triangulate) {
-      size_t k;
-      size_t n = 0;
+      unsigned int k;
+      unsigned int n = 0;
 
       tinyobj_vertex_index_t i0 = f[0];
       tinyobj_vertex_index_t i1;
@@ -1113,7 +1113,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
       command->num_f_num_verts = n;
 
     } else {
-      size_t k = 0;
+      unsigned int k = 0;
       assert(num_f < TINYOBJ_MAX_FACES_PER_F_LINE);
       for (k = 0; k < num_f; k++) {
         command->f[k] = f[k];
@@ -1134,7 +1134,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
     skip_space(&token);
     command->material_name = p + (token - linebuf);
     command->material_name_len = (unsigned int)length_until_newline(
-                                                                    token, (p_len - (size_t)(token - linebuf)) + 1);
+                                                                    token, (p_len - (unsigned int)(token - linebuf)) + 1);
     command->type = COMMAND_USEMTL;
 
     return 1;
@@ -1148,7 +1148,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
     skip_space(&token);
     command->mtllib_name = p + (token - linebuf);
     command->mtllib_name_len = (unsigned int)length_until_newline(
-                                                                  token, p_len - (size_t)(token - linebuf)) +
+                                                                  token, p_len - (unsigned int)(token - linebuf)) +
       1;
     command->type = COMMAND_MTLLIB;
 
@@ -1162,7 +1162,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
 
     command->group_name = p + (token - linebuf);
     command->group_name_len = (unsigned int)length_until_newline(
-                                                                 token, p_len - (size_t)(token - linebuf)) +
+                                                                 token, p_len - (unsigned int)(token - linebuf)) +
       1;
     command->type = COMMAND_G;
 
@@ -1176,7 +1176,7 @@ static int parseLine(Command *command, const char *p, size_t p_len,
 
     command->object_name = p + (token - linebuf);
     command->object_name_len = (unsigned int)length_until_newline(
-                                                                  token, p_len - (size_t)(token - linebuf)) +
+                                                                  token, p_len - (unsigned int)(token - linebuf)) +
       1;
     command->type = COMMAND_O;
 
@@ -1187,11 +1187,11 @@ static int parseLine(Command *command, const char *p, size_t p_len,
 }
 
 typedef struct {
-  size_t pos;
-  size_t len;
+  unsigned int pos;
+  unsigned int len;
 } LineInfo;
 
-static int is_line_ending(const char *p, size_t i, size_t end_i) {
+static int is_line_ending(const char *p, unsigned int i, unsigned int end_i) {
   if (p[i] == '\0') return 1;
   if (p[i] == '\n') return 1; /* this includes \r\n */
   if (p[i] == '\r') {
@@ -1203,23 +1203,23 @@ static int is_line_ending(const char *p, size_t i, size_t end_i) {
 }
 
 int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
-                      size_t *num_shapes, tinyobj_material_t **materials_out,
-                      size_t *num_materials_out, const char *buf, size_t len,
+                      unsigned int *num_shapes, tinyobj_material_t **materials_out,
+                      unsigned int *num_materials_out, const char *buf, unsigned int len,
                       unsigned int flags) {
   LineInfo *line_infos = NULL;
   Command *commands = NULL;
-  size_t num_lines = 0;
+  unsigned int num_lines = 0;
 
-  size_t num_v = 0;
-  size_t num_vn = 0;
-  size_t num_vt = 0;
-  size_t num_f = 0;
-  size_t num_faces = 0;
+  unsigned int num_v = 0;
+  unsigned int num_vn = 0;
+  unsigned int num_vt = 0;
+  unsigned int num_f = 0;
+  unsigned int num_faces = 0;
 
   int mtllib_line_index = -1;
 
   tinyobj_material_t *materials = NULL;
-  size_t num_materials = 0;
+  unsigned int num_materials = 0;
 
   hash_table_t material_table;
 
@@ -1234,11 +1234,11 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
   tinyobj_attrib_init(attrib);
    /* 1. Find '\n' and create line data. */
   {
-    size_t i;
-    size_t end_idx = len;
-    size_t prev_pos = 0;
-    size_t line_no = 0;
-    size_t last_line_ending = 0;
+    unsigned int i;
+    unsigned int end_idx = len;
+    unsigned int prev_pos = 0;
+    unsigned int line_no = 0;
+    unsigned int last_line_ending = 0;
 
     /* Count # of lines. */
     for (i = 0; i < end_idx; i++) {
@@ -1280,7 +1280,7 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
 
   /* 2. parse each line */
   {
-    size_t i = 0;
+    unsigned int i = 0;
     for (i = 0; i < num_lines; i++) {
       int ret = parseLine(&commands[i], &buf[line_infos[i].pos],
                           line_infos[i].len, flags & TINYOBJ_FLAG_TRIANGULATE);
@@ -1328,13 +1328,13 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
   /* Construct attributes */
 
   {
-    size_t v_count = 0;
-    size_t n_count = 0;
-    size_t t_count = 0;
-    size_t f_count = 0;
-    size_t face_count = 0;
+    unsigned int v_count = 0;
+    unsigned int n_count = 0;
+    unsigned int t_count = 0;
+    unsigned int f_count = 0;
+    unsigned int face_count = 0;
     int material_id = -1; /* -1 = default unknown material. */
-    size_t i = 0;
+    unsigned int i = 0;
 
     attrib->vertices = (float *)TINYOBJ_MALLOC(sizeof(float) * num_v * 3);
     attrib->num_vertices = (unsigned int)num_v;
@@ -1398,7 +1398,7 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
         attrib->texcoords[2 * t_count + 1] = commands[i].ty;
         t_count++;
       } else if (commands[i].type == COMMAND_F) {
-        size_t k = 0;
+        unsigned int k = 0;
         for (k = 0; k < commands[i].num_f; k++) {
           tinyobj_vertex_index_t vi = commands[i].f[k];
           int v_idx = fixIndex(vi.v_idx, v_count);
@@ -1423,9 +1423,9 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
   /* 5. Construct shape information. */
   {
     unsigned int face_count = 0;
-    size_t i = 0;
-    size_t n = 0;
-    size_t shape_idx = 0;
+    unsigned int i = 0;
+    unsigned int n = 0;
+    unsigned int shape_idx = 0;
 
     const char *shape_name = NULL;
     unsigned int shape_name_len = 0;
@@ -1497,7 +1497,7 @@ int tinyobj_parse_obj(tinyobj_attrib_t *attrib, tinyobj_shape_t **shapes,
     }
 
     if ((face_count - prev_face_offset) > 0) {
-      size_t length = face_count - prev_shape_face_offset;
+      unsigned int length = face_count - prev_shape_face_offset;
       if (length > 0) {
         (*shapes)[shape_idx].name =
           my_strndup(prev_shape_name, prev_shape_name_len);
@@ -1548,8 +1548,8 @@ void tinyobj_attrib_free(tinyobj_attrib_t *attrib) {
   if (attrib->material_ids) TINYOBJ_FREE(attrib->material_ids);
 }
 
-void tinyobj_shapes_free(tinyobj_shape_t *shapes, size_t num_shapes) {
-  size_t i;
+void tinyobj_shapes_free(tinyobj_shape_t *shapes, unsigned int num_shapes) {
+  unsigned int i;
   if (shapes == NULL) return;
 
   for (i = 0; i < num_shapes; i++) {
@@ -1560,8 +1560,8 @@ void tinyobj_shapes_free(tinyobj_shape_t *shapes, size_t num_shapes) {
 }
 
 void tinyobj_materials_free(tinyobj_material_t *materials,
-                            size_t num_materials) {
-  size_t i;
+                            unsigned int num_materials) {
+  unsigned int i;
   if (materials == NULL) return;
 
   for (i = 0; i < num_materials; i++) {
